% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{continuous_variables_in_data}
\alias{continuous_variables_in_data}
\title{Get continuous variables from data.frame}
\usage{
continuous_variables_in_data(data_df)
}
\arguments{
\item{data}{data.frame - data to look through}
}
\value{
vector of column names
}
\description{
\code{continuous_variables_in_data} helper function to pull out the column names of continuous variables in a data.frame
}
\details{
Takes a data.frame and check the columns to see which are continuous.
   Returns a vector of column names
}
\examples{
continuous_variables_in_data( data_frame )

}
