% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{factors_in_data}
\alias{factors_in_data}
\title{Get factors from a data.frame}
\usage{
factors_in_data(data_df)
}
\arguments{
\item{data}{data.frame - data to look through}
}
\value{
vector of column names
}
\description{
\code{factors_in_data} helper function to pull out the column names of factors in a data.frame
}
\details{
Takes a data.frame and check the columns to see which is a factor.
   Returns a vector of column names
}
\examples{
factors_in_data( data_frame )

}
