% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_app.R
\name{fileUploadUI}
\alias{fileUploadUI}
\title{File Upload UI}
\usage{
fileUploadUI(id, label = "File")
}
\arguments{
\item{id}{A unique label for the module}

\item{label}{A label to show on the file input panel.}
}
\value{
A sharable file upload module
}
\description{
File Upload UI
}
\examples{
\dontrun{
fileUploadUI(id = "my_file_upload" , label = "Input File")
}
}
