% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_app.R
\name{modular_app}
\alias{modular_app}
\title{Run an example module based app}
\usage{
modular_app()
}
\value{
Returns an example module based app.
}
\description{
Run an example module based app
}
\examples{
\dontrun{
run_app("modular_app")
}
}
