% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_modules.R
\name{more_modules}
\alias{more_modules}
\title{Run another module based app}
\usage{
more_modules()
}
\value{
Returns an example module based app.
}
\description{
Run another module based app
}
\examples{
\dontrun{
run_app("more_modules")
}
}
