% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_app.R
\name{ReadFileServer}
\alias{ReadFileServer}
\title{File Reader Uploader}
\usage{
ReadFileServer(id)
}
\arguments{
\item{id}{A unique id for the file reader server}
}
\value{
A server module to read CSV files from fileUploadUI
}
\description{
File Reader Uploader
}
\examples{
\dontrun{
ReadFileServer("my_reader_module")
}
}
