% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run apps in shinymisc}
\usage{
run_app(app_name = "theming")
}
\arguments{
\item{app_name}{Name of the app to run}
}
\value{
A shiny app
}
\description{
Run apps in shinymisc
}
\examples{
if (interactive()) run_app("theming")
}
