% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{theming_app}
\alias{theming_app}
\title{An app to demonstrate theming of a plot in shiny}
\usage{
theming_app()
}
\description{
An app to demonstrate theming of a plot in shiny
}
\examples{
if(interactive()) theming_app()
}
