% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{themingServer}
\alias{themingServer}
\title{Server for the theming app}
\usage{
themingServer(input, output, session)
}
\arguments{
\item{input}{input from UI}

\item{output}{output server}

\item{session}{session}
}
\value{
a server module
}
\description{
Server for the theming app
}
