% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{themingServer}
\alias{themingServer}
\title{Server for the theming app}
\usage{
themingServer(id)
}
\arguments{
\item{id}{An id for use with moduleServer}
}
\value{
a server module
}
\description{
Server for the theming app
}
